Adding an instance for the +Val+ classtype for values of the datatype +Maybe+.
Reasoning:

Most of the times, when we want to work with null values in Haskell we use the +Maybe+ datatype. In this library we are using the  +Null+ constructor of the datatype +Value+ defined in the Bson module.

When using this on an insert, We get the following code:

    insert "posts" [u"title"   =: u"My inspiring blogpost."
                   ,u"content" =: u"This is my first blogpost."
                   ,u"date"   := Null]

There are two problems regarding the Null approach:
  - Haskellers have to know the constructors +Null+ of the datatype +Value+ in order to get to this conclusion, instead of using the more intuitive (at least for Haskellers) +Nothing+ value
  - We are using another operator (:=) to create the +Field+ value instead of the (=:) operator, this type of inconsistencies tend to create "WTF moments" on the code.

With the small implementation in this patch, we are able to do the following:

    insert "posts" [u"title"   =: u"My inspiring blogpost."
                   ,u"content" =: u"This is my first blogpost."
                   ,u"date"    =: (Nothing :: Maybe UTCTime)]

Although this solution is a bit more verbose (we have to state the type on the signature), is a more intuitive solution for middle experienced Haskellers. In the same way, it uses the same operator (=:) consistently on all the fields.